+++ src/epar/util/SymbolPool.java
@@ -5,15 +5,15 @@ import java.util.Map;
 
 public class SymbolPool {
 
-    private static final Map<String, Short> stringToID = new HashMap<>();
+    private static final Map<String, Integer> stringToID = new HashMap<>();
 
-    private static final Map<Short, String> idToString = new HashMap<>();
+    private static final Map<Integer, String> idToString = new HashMap<>();
     
     public static final short NONE = 0;
 
-    private static short nextID = 1;
+    private static int nextID = 1;
 
-    public static short getID(String string) {
+    public static int getID(String string) {
         if (!stringToID.containsKey(string)) {
             if (nextID == 0) {
                 throw new RuntimeException("Symbol pool overflow!");
@@ -27,7 +27,7 @@ public class SymbolPool {
         return stringToID.get(string);
     }
 
-    public static String getString(short id) {
+    public static String getString(int id) {
         return idToString.get(id);
     }
